rename and move files, update junit #25
