+++ statique/pom.xml
@@ -12,7 +12,7 @@
     <maven.compiler.source>11</maven.compiler.source>
     <maven.compiler.target>11</maven.compiler.target>
     <picocli.version>4.6.1</picocli.version>
-    <junit.version>4.11</junit.version>
+    <junit.version>4.13.1</junit.version>
     <maven-assembly-plugin.version>3.3.0</maven-assembly-plugin.version>
   </properties>
 
@@ -32,7 +32,7 @@
               <archive>
                 <manifest>
                   <mainClass>
-                    com.mycompany.app.Main
+                    ch.heigvd.igjt.statique.Main
                   </mainClass>
                 </manifest>
               </archive>
