+++ statique/src/main/java/ch/heigvd/igjt/statique/subcommands/SubCommandBuild.java
@@ -1,4 +1,4 @@
-package ch.heigvd.igjt.subcommands.statique;
+package ch.heigvd.igjt.statique.subcommands;
 
 import picocli.CommandLine;
