+++ statique/src/main/java/ch/heigvd/igjt/statique/subcommands/SubCommandNew.java
@@ -1,4 +1,4 @@
-package ch.heigvd.igjt.subcommands;
+package ch.heigvd.igjt.statique.subcommands;
 
 import picocli.CommandLine;
